+++ docs/changelog.md
@@ -20,6 +20,7 @@ pages](https://github.com/dalibo/ldap2pg/pulls?utf8=%E2%9C%93&q=is%3Apr%20is%3Am
 - Reassign database before dropping owner.
 - Reassign objects to database owner instead of managing role.
 - Configure fallback owner, default to managing role.
+- Set application_name to ldap2pg.
 - Configure SASL_MECH from YAML.
 - Build RHEL 8 package with Rockylinux 8
 - Move RHEL 7 package to Python 3.6.